Rework matchAccountID to be findMatchedProperty.
