+++ assets/js/modules/analytics-4/datastore/properties.js
@@ -173,6 +173,34 @@ const baseActions = {
 	),
 
 	/**
+	 * Gets an ID of an account that has a matching property.
+	 *
+	 * @since n.e.x.t
+	 *
+	 * @return {string} The account ID on success, otherwise NULL.
+	 */
+	*findMatchedProperty() {
+		const registry = yield commonActions.getRegistry();
+		const accounts = yield Data.commonActions.await(
+			registry.__experimentalResolveSelect( STORE_NAME ).getAccountSummaries()
+		);
+
+		if ( ! Array.isArray( accounts ) || accounts.length === 0 ) {
+			return null;
+		}
+
+		const url = registry.select( CORE_SITE ).getReferenceSiteURL();
+		const propertyIDs = accounts.reduce(
+			( acc, { propertySummaries: properties } ) => [ ...acc, ...( properties || [] ).map( ( { _id } ) => _id ) ],
+			[],
+		);
+
+		return yield Data.commonActions.await(
+			registry.dispatch( STORE_NAME ).matchPropertyByURL( propertyIDs, url )
+		);
+	},
+
+	/**
 	 * Matches and selects a property for provided accountID.
 	 *
 	 * @since 1.34.0
